create a generateClass function
