+++ js/vault.js
@@ -1,5 +1,5 @@
-const VaultAbi = require("../build/Vault.sol").VaultAbi;
-const VaultByteCode = require("../build/Vault.sol").VaultByteCode;
-const runethtx = require("runethtx");
+const VaultAbi = require('../build/Vault.sol').VaultAbi;
+const VaultByteCode = require('../build/Vault.sol').VaultByteCode;
+const generateClass = require('./generateClass');
 
-module.exports = runethtx.generateClass(VaultAbi, VaultByteCode);
+module.exports = generateClass(VaultAbi, VaultByteCode);
